+++ app/src/main/java/com/askfortricks/mvvmjetpackdemo/model/ApiResponseModels.kt
@@ -1,4 +1,4 @@
-package com.askfortricks.mvvmjetpackdemo.data_models
+package com.askfortricks.mvvmjetpackdemo.model
 
 
 data class MovieList(
