+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/movieListFragment"
-        android:name="com.askfortricks.mvvmjetpackdemo.fragments.MovieListFragment"
+        android:name="com.askfortricks.mvvmjetpackdemo.view.MovieListFragment"
         android:label="Movie List"
         tools:layout="@layout/fragment_movie_list">
         <action
@@ -16,7 +16,7 @@
     </fragment>
     <fragment
         android:id="@+id/movieDetailFragment"
-        android:name="com.askfortricks.mvvmjetpackdemo.fragments.MovieDetailFragment"
+        android:name="com.askfortricks.mvvmjetpackdemo.view.MovieDetailFragment"
         android:label="MovieDetailFragment"
         tools:layout="@layout/fragment_movie_detail">
     </fragment>
